pdbdom.c: Cut runtime in half, memory consumption by 30%
This is done by getting rid of the far array, storing its contents
temporarily in the PDB instead.  This improves both memory usage
(one less huge array) and runtime as we only need to do one lookup
instead of a binary search to find if a configuration has already
been dominated.