puzzle.c: Fix transpose() function, add transpositions[]
In the previous implementation, transpose essentially computed

    tiles = tiles * transpositions,
    grid  = transpositions * grid.

However, while transposing the puzzle configuration, this map is not
the map we want as it does not preserve the distance to the solved
configuration.  This is only achieved by the map

    tiles = transposition * tiles * transposition
    grid  = transposition * grid  * transposition

where transposition works like a change of basis.  Note that we do not
need to invert transposition as it is an involution.

This commit also deletes all the test results that are now invalid.