pdb.c: Simplify PDB structure
We can allocate the whole PDB in one chunk, this saves one read from poorly
cached memory for an extra (cheap) multiplication.  Sadly, pdb_ident() is
a bit more difficult to implement using this new scheme, so it is removed
for the time being.  The PDB layout has not changed, so old identified PDBs
can still be used.